Implement L2Config for ProverClientConfig. Move SequencerConfig to new file
