+++ crates/l2/utils/config/toml_parser.rs
@@ -2,40 +2,8 @@ use crate::utils::config::{
     errors::{ConfigError, TomlParserError},
     ConfigMode,
 };
-use serde::Deserialize;
 
-use super::{
-    block_producer::BlockProducerConfig, committer::CommitterConfig, deployer::DeployerConfig,
-    eth::EthConfig, l1_watcher::L1WatcherConfig, prover_client::ProverClientConfig,
-    prover_server::ProverServerConfig, L2Config,
-};
-
-#[derive(Deserialize, Debug)]
-pub struct SequencerConfig {
-    deployer: DeployerConfig,
-    eth: EthConfig,
-    watcher: L1WatcherConfig,
-    proposer: BlockProducerConfig,
-    committer: CommitterConfig,
-    prover_server: ProverServerConfig,
-}
-
-impl L2Config for SequencerConfig {
-    const PREFIX: &str = "";
-
-    fn to_env(&self) -> String {
-        let mut env_representation = String::new();
-
-        env_representation.push_str(&self.deployer.to_env());
-        env_representation.push_str(&self.eth.to_env());
-        env_representation.push_str(&self.watcher.to_env());
-        env_representation.push_str(&self.proposer.to_env());
-        env_representation.push_str(&self.committer.to_env());
-        env_representation.push_str(&self.prover_server.to_env());
-
-        env_representation
-    }
-}
+use super::{prover_client::ProverClientConfig, sequencer::SequencerConfig, L2Config};
 
 fn read_config(config_path: String, mode: ConfigMode) -> Result<(), ConfigError> {
     let toml_path = mode.get_config_file_path(&config_path);
@@ -53,13 +21,13 @@ fn read_config(config_path: String, mode: ConfigMode) -> Result<(), ConfigError>
             let config: SequencerConfig = toml::from_str(&file).map_err(|err| {
                 TomlParserError::TomlFormat(format!("{err}: {}", toml_file_name.clone()), mode)
             })?;
-            config.write_env()?;
+            config.write_env(&mode.get_env_path_or_default())?;
         }
         ConfigMode::ProverClient => {
             let config: ProverClientConfig = toml::from_str(&file).map_err(|err| {
                 TomlParserError::TomlFormat(format!("{err}: {}", toml_file_name.clone()), mode)
             })?;
-            config.write_env()?;
+            config.write_env(&mode.get_env_path_or_default())?;
         }
     }
